move extensions to correct files
