+++ lib/modules/manga/reader/providers/reader_controller_provider.dart
@@ -1,21 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_riverpod/misc.dart';
-import 'package:isar_community/isar.dart';
 import 'package:mangayomi/main.dart';
 import 'package:mangayomi/models/chapter.dart';
-import 'package:mangayomi/models/download.dart';
 import 'package:mangayomi/modules/manga/reader/mixins/chapter_reader_settings_mixin.dart';
 import 'package:mangayomi/modules/manga/reader/mixins/chapter_controller_mixin.dart';
-import 'package:mangayomi/models/manga.dart';
 import 'package:mangayomi/models/settings.dart';
-import 'package:mangayomi/models/track.dart';
-import 'package:mangayomi/models/track_preference.dart';
-import 'package:mangayomi/modules/manga/detail/providers/track_state_providers.dart';
 import 'package:mangayomi/modules/more/providers/incognito_mode_state_provider.dart';
 import 'package:mangayomi/modules/more/settings/downloads/providers/downloads_state_provider.dart';
-import 'package:mangayomi/modules/more/settings/track/providers/track_providers.dart';
-import 'package:mangayomi/utils/chapter_recognition.dart';
 import 'package:mangayomi/utils/extensions/chapter.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 part 'reader_controller_provider.g.dart';
@@ -234,176 +225,4 @@ class ReaderController extends _$ReaderController
       }
     }
   }
-}
-
-extension ChapterExtensions on Chapter {
-  void updateTrackChapterRead(dynamic ref) {
-    if (!(ref is WidgetRef || ref is Ref)) return;
-    final updateProgressAfterReading = ref.read(
-      updateProgressAfterReadingStateProvider,
-    );
-    if (!updateProgressAfterReading) return;
-    final manga = this.manga.value!;
-    final chapterNumber = ChapterRecognition().parseChapterNumber(
-      manga.name!,
-      name!,
-    );
-
-    final tracks = isar.tracks
-        .filter()
-        .idIsNotNull()
-        .itemTypeEqualTo(manga.itemType)
-        .mangaIdEqualTo(manga.id!)
-        .findAllSync();
-
-    if (tracks.isEmpty) return;
-    for (var track in tracks) {
-      final service = isar.trackPreferences
-          .filter()
-          .syncIdIsNotNull()
-          .syncIdEqualTo(track.syncId)
-          .findFirstSync();
-      if (!(service == null || chapterNumber <= (track.lastChapterRead ?? 0))) {
-        if (track.status != TrackStatus.completed) {
-          track.lastChapterRead = chapterNumber;
-          if (track.lastChapterRead == track.totalChapter &&
-              (track.totalChapter ?? 0) > 0) {
-            track.status = TrackStatus.completed;
-            track.finishedReadingDate = DateTime.now().millisecondsSinceEpoch;
-          } else {
-            track.status = manga.itemType == ItemType.manga
-                ? TrackStatus.reading
-                : TrackStatus.watching;
-            if (track.lastChapterRead == 1) {
-              track.startedReadingDate = DateTime.now().millisecondsSinceEpoch;
-            }
-          }
-        }
-        ref
-            .read(
-              trackStateProvider(
-                track: track,
-                itemType: manga.itemType,
-                widgetRef: ref,
-              ).notifier,
-            )
-            .updateManga();
-      }
-    }
-  }
-}
-
-extension MangaExtensions on Manga {
-  // ── For the READER: always ascending story order, filters applied ──────────
-  List<Chapter> getChapterListForReading() {
-    final settings = isar.settings.getSync(227)!;
-
-    final filterUnread =
-        (settings.chapterFilterUnreadList!
-                    .where((e) => e.mangaId == id)
-                    .firstOrNull ??
-                ChapterFilterUnread(mangaId: id, type: 0))
-            .type!;
-
-    final filterBookmarked =
-        (settings.chapterFilterBookmarkedList!
-                    .where((e) => e.mangaId == id)
-                    .firstOrNull ??
-                ChapterFilterBookmarked(mangaId: id, type: 0))
-            .type!;
-
-    final filterDownloaded =
-        (settings.chapterFilterDownloadedList!
-                    .where((e) => e.mangaId == id)
-                    .firstOrNull ??
-                ChapterFilterDownloaded(mangaId: id, type: 0))
-            .type!;
-
-    final filterScanlator = _getFilterScanlator(this) ?? [];
-
-    // Canonical ascending order (ch1 ... chN) — reader always moves forward.
-    final data = chapters
-        .toList(); // keep DB/insertion order, assumed ascending
-
-    final chapterIds = data.map((c) => c.id).whereType<int>().toList();
-    final downloadedIds = (filterDownloaded == 0 || chapterIds.isEmpty)
-        ? const <int>{}
-        : isar.downloads
-              .filter()
-              .anyOf(chapterIds, (q, id) => q.idEqualTo(id))
-              .isDownloadEqualTo(true)
-              .findAllSync()
-              .map((d) => d.id!)
-              .toSet();
-
-    return data
-        .where(
-          (e) => filterUnread == 1
-              ? e.isRead == false
-              : filterUnread == 2
-              ? e.isRead == true
-              : true,
-        )
-        .where(
-          (e) => filterBookmarked == 1
-              ? e.isBookmarked == true
-              : filterBookmarked == 2
-              ? e.isBookmarked == false
-              : true,
-        )
-        .where((e) {
-          if (filterDownloaded == 0) return true;
-          final dl = downloadedIds.contains(e.id);
-          return filterDownloaded == 1 ? dl : !dl;
-        })
-        .where((e) => !filterScanlator.contains(e.scanlator))
-        .toList();
-  }
-
-  // ── For the UI LIST: filters + user-chosen sort + reverse ─────────────────
-  List<Chapter> getFilteredChapterList() {
-    final settings = isar.settings.getSync(227)!;
-
-    final sortChapterEntry =
-        settings.sortChapterList!.where((e) => e.mangaId == id).firstOrNull ??
-        SortChapter(mangaId: id, index: 1, reverse: false);
-    final sortIndex = sortChapterEntry.index!;
-    final reverse = sortChapterEntry.reverse!;
-
-    // Start from the reading list so filter logic lives in one place.
-    List<Chapter> list = getChapterListForReading();
-
-    switch (sortIndex) {
-      case 0: // by scanlator, then date
-        list.sort((a, b) {
-          if (a.scanlator == null || b.scanlator == null) return 0;
-          final s = a.scanlator!.compareTo(b.scanlator!);
-          if (s != 0) return s;
-          if (a.dateUpload == null || b.dateUpload == null) return 0;
-          return int.parse(a.dateUpload!).compareTo(int.parse(b.dateUpload!));
-        });
-      case 1: // by chapter number - reading list is already ascending
-        break;
-      case 2: // by upload date
-        list.sort((a, b) {
-          if (a.dateUpload == null || b.dateUpload == null) return 0;
-          return int.parse(a.dateUpload!).compareTo(int.parse(b.dateUpload!));
-        });
-      case 3: // by name
-        list.sort((a, b) {
-          if (a.name == null || b.name == null) return 0;
-          return a.name!.compareTo(b.name!);
-        });
-    }
-
-    return reverse ? list.reversed.toList() : list;
-  }
-}
-
-List<String>? _getFilterScanlator(Manga manga) {
-  final scanlators = isar.settings.getSync(227)!.filterScanlatorList ?? [];
-  final filter = scanlators
-      .where((element) => element.mangaId == manga.id)
-      .toList();
-  return filter.firstOrNull?.scanlators;
 }
